segunda-alt

+++ 21.cpp
@@ -7,7 +7,7 @@ float M, K;
 cout<<"Qual a velocidade em m/s? \n";
 cin>>M;
 
-K=M*3.6;
+K=(double)M*3.6;
 cout<<"A velocidade em k/h e "<<K;
 
 return 0;
